Testcontainers 통합 테스트 추가 및 스키마-엔티티 타입 불일치 수정
주문 생성(MS-16)과 메뉴 시드가 실제 MySQL 8.4 컨테이너에서 저장·검증까지 되는지
확인하는 통합 테스트를 추가했다. 이 과정에서 JPA 엔티티 필드 타입이 실제 DB
물리 타입(TINYINT/SMALLINT/CHAR)과 어긋나 스키마 검증이 깨지는 버그 3건을
찾아 고쳤다 — Mockito/H2로는 잡히지 않고 실제 MySQL에서만 드러나는 종류였다.

- OptionGroup.minSelect/maxSelect: int → byte (TINYINT)
- OrderLine.quantity: int → short (SMALLINT)
- Order.requestHash: columnDefinition="CHAR(64)" 명시 (기본 VARCHAR 추론 방지)

테스트 컨테이너는 클래스별로 띄우지 않고 전체 실행에서 하나만 공유하는
싱글턴 패턴으로 구성했다 (클래스마다 새로 띄우면 로컬 Colima VM 메모리로
두 개를 동시에 못 버팀). testcontainers-bom을 1.20.4 → 1.21.4로 올려
최신 Docker 엔진(29.x)과의 API 버전 협상 실패도 해결했다.

+++ src/main/java/com/coffeul/menu/domain/OptionGroup.java
@@ -29,10 +29,10 @@ public class OptionGroup {
     private boolean required = true;
 
     @Column(name = "min_select", nullable = false)
-    private int minSelect = 1;
+    private byte minSelect = 1;
 
     @Column(name = "max_select", nullable = false)
-    private int maxSelect = 1;
+    private byte maxSelect = 1;
 
     @Column(name = "sort_order", nullable = false)
     private int sortOrder;
